Remove dependencies from tests/e2e/framework on pkg/scheduler/framework
