+++ pkg/scheduler/framework/v1alpha1/framework_test.go
@@ -1763,6 +1763,58 @@ func TestListPlugins(t *testing.T) {
 	}
 }
 
+func TestDefaultNormalizeScore(t *testing.T) {
+	tests := []struct {
+		reverse        bool
+		scores         []int64
+		expectedScores []int64
+	}{
+		{
+			scores:         []int64{1, 2, 3, 4},
+			expectedScores: []int64{25, 50, 75, 100},
+		},
+		{
+			reverse:        true,
+			scores:         []int64{1, 2, 3, 4},
+			expectedScores: []int64{75, 50, 25, 0},
+		},
+		{
+			scores:         []int64{1000, 10, 20, 30},
+			expectedScores: []int64{100, 1, 2, 3},
+		},
+		{
+			reverse:        true,
+			scores:         []int64{1000, 10, 20, 30},
+			expectedScores: []int64{0, 99, 98, 97},
+		},
+		{
+			scores:         []int64{1, 1, 1, 1},
+			expectedScores: []int64{100, 100, 100, 100},
+		},
+		{
+			scores:         []int64{1000, 1, 1, 1},
+			expectedScores: []int64{100, 0, 0, 0},
+		},
+	}
+
+	for i, test := range tests {
+		scores := NodeScoreList{}
+		for _, score := range test.scores {
+			scores = append(scores, NodeScore{Score: score})
+		}
+
+		expectedScores := NodeScoreList{}
+		for _, score := range test.expectedScores {
+			expectedScores = append(expectedScores, NodeScore{Score: score})
+		}
+
+		DefaultNormalizeScore(MaxNodeScore, test.reverse, scores)
+		if !reflect.DeepEqual(scores, expectedScores) {
+			t.Errorf("test %d, expected %v, got %v", i, expectedScores, scores)
+		}
+	}
+}
+
 func buildScoreConfigDefaultWeights(ps ...string) *config.Plugins {
 	return buildScoreConfigWithWeights(defaultWeights, ps...)
 }
