+++ test/e2e/framework/.import-restrictions
@@ -113,7 +113,6 @@
 				"k8s.io/kubernetes/pkg/kubelet/kuberuntime",
 				"k8s.io/kubernetes/pkg/kubelet/kuberuntime/logs",
 				"k8s.io/kubernetes/pkg/kubelet/leaky",
-				"k8s.io/kubernetes/pkg/kubelet/lifecycle",
 				"k8s.io/kubernetes/pkg/kubelet/logs",
 				"k8s.io/kubernetes/pkg/kubelet/metrics",
 				"k8s.io/kubernetes/pkg/kubelet/network/dns",
@@ -146,7 +145,6 @@
 				"k8s.io/kubernetes/pkg/kubelet/stats/pidlimit",
 				"k8s.io/kubernetes/pkg/kubelet/status",
 				"k8s.io/kubernetes/pkg/kubelet/secret",
-				"k8s.io/kubernetes/pkg/kubelet/sysctl",
 				"k8s.io/kubernetes/pkg/kubelet/types",
 				"k8s.io/kubernetes/pkg/kubelet/token",
 				"k8s.io/kubernetes/pkg/kubelet/util",
@@ -180,7 +178,7 @@
 				"k8s.io/kubernetes/pkg/proxy/util",
 				"k8s.io/kubernetes/pkg/registry/core/service/allocator",
 				"k8s.io/kubernetes/pkg/registry/core/service/portallocator",
-				"k8s.io/kubernetes/pkg/scheduler/api",
+				"k8s.io/kubernetes/pkg/scheduler/apis/config",
 				"k8s.io/kubernetes/pkg/scheduler/framework/plugins/helper",
 				"k8s.io/kubernetes/pkg/scheduler/framework/plugins/nodeaffinity",
 				"k8s.io/kubernetes/pkg/scheduler/framework/plugins/nodename",
@@ -192,7 +190,6 @@
 				"k8s.io/kubernetes/pkg/scheduler/metrics",
 				"k8s.io/kubernetes/pkg/scheduler/nodeinfo",
 				"k8s.io/kubernetes/pkg/scheduler/util",
-				"k8s.io/kubernetes/pkg/scheduler/volumebinder",
 				"k8s.io/kubernetes/pkg/security/apparmor",
 				"k8s.io/kubernetes/pkg/security/podsecuritypolicy/seccomp",
 				"k8s.io/kubernetes/pkg/security/podsecuritypolicy/sysctl",
